[Refactor] base 패키지 안에 불필요한 패키지 삭제
