--- app/src/main/java/com/doctor/yumyum/common/base/adapter/BaseRecyclerAdapter.kt
@@ -1,5 +1,0 @@
-package com.doctor.yumyum.common.base.adapter
-
-abstract class BaseRecyclerAdapter{
-
-}
