+++ app/src/main/java/com/doctor/yumyum/presentation/ui/MainActivity.kt
@@ -3,7 +3,7 @@ package com.doctor.yumyum.presentation.ui
 import android.os.Bundle
 import com.doctor.yumyum.R
 import com.doctor.yumyum.databinding.ActivityMainBinding
-import com.doctor.yumyum.common.base.activity.BaseActivity
+import com.doctor.yumyum.common.base.BaseActivity
 
 class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
     override fun onCreate(savedInstanceState: Bundle?) {
