+++ app/src/main/java/com/doctor/yumyum/presentation/viewmodel/MainViewModel.kt
@@ -1,6 +1,6 @@
 package com.doctor.yumyum.presentation.viewmodel
 
-import com.doctor.yumyum.common.base.viewmodel.BaseViewModel
+import com.doctor.yumyum.common.base.BaseViewModel
 
 /** viewModel package 임시저장용 나중에 삭제예정 **/
 class MainViewModel : BaseViewModel() {
